+++ backend/src/delivery/domain/repository/trade_partner_reposiory_trait.rs
@@ -1,13 +1,13 @@
 use crate::delivery::domain::TradePartner;
 
 pub trait TradePartnerTrait<T> {
-    fn insert(db_pool: T, trade_partner: TradePartner) -> Option<usize>;
+    fn insert(db_pool: T, trade_partner: TradePartner) -> Option<i32>;
 
     fn save(tdb_pool: T, rade_partner: TradePartner) -> bool;
 
-    fn delete(db_pool: T, id: usize) -> bool;
+    fn delete(db_pool: T, id: i32) -> bool;
 
-    fn find_by_id(db_pool: T, id: usize) -> Option<TradePartner>;
+    fn find_by_id(db_pool: T, id: i32) -> Option<TradePartner>;
 
     fn get_all(db_pool: T) -> Vec<TradePartner>;
 }
