+++ backend/src/delivery/domain/warehouse.rs
@@ -1,13 +1,13 @@
 #[derive(Clone)]
 pub struct Warehouse {
-    pub id: usize,
+    pub id: i32,
     pub name: String,
-    pub trade_partner_id: usize,
-    pub address_id: usize,
+    pub trade_partner_id: i32,
+    pub address_id: i32,
 }
 
 impl Warehouse {
-    pub fn new(name: String, trade_partner_id: usize, address_id: usize) -> Self {
+    pub fn new(name: String, trade_partner_id: i32, address_id: i32) -> Self {
         Warehouse {
             id: 0,
             name,
